+++ src/js/block-editor.js
@@ -12,7 +12,7 @@ import { registerBlockCollection } from '@wordpress/blocks';
  * @link https://github.com/godaddy-wordpress/coblocks/blob/e196c83d628e1139799ebf9ef94540578f430141/src/utils/helper.js
  */
 registerBlockCollection( 'r3-id-documentation', {
-	title: 'r3-id-documentation Collection (JS)',
+	title: 'r3-id-documentation Blocks Collection (JS)',
 } );
 
 /**