+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # Changelog
 
-## 0.1.0
+## FORTHCOMING
 
 - Added Embed Markdown block
+
+## 0.1.0
+
+- Adds `<LoadingSpinner>` sandbox block
+- Adds tutorial-01 block for documentation
 
 ## 0.0.0